Refactor all of the annotation processing for SourceType into SourceUtil
